+++ app/src/main/java/com/macochave/tupperinventario/datos/tad/TADColor.java
@@ -6,22 +6,22 @@ package com.macochave.tupperinventario.datos.tad;
 
 public class TADColor {
 
-    private int id;
+    private long id;
     private String color;
 
     public TADColor() {
     }
 
-    public TADColor(int id, String color) {
+    public TADColor(long id, String color) {
         this.id = id;
         this.color = color;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
